Move parse_command_line_arguments to util.py
rmg.py is not installed as a module in the conda binary, so we cannot import this function
